refactor: use one single octokit module
Also works with dotenv. Untested.
