+++ src/dotenv.mjs
@@ -1,0 +1,4 @@
+import dotenv from 'dotenv'
+
+// Read the `.env` file in the source root dir if it exists
+dotenv.config()
