+++ test/e2e-test.mjs
@@ -1,20 +1,11 @@
-import { Octokit } from '@octokit/rest'
-import { graphql } from '@octokit/graphql'
-import generateReleaseNotes from '../src/generate-release-notes.mjs'
+import getOctokit from '../src/octokit.mjs'
+import { lookupLoginIds } from '../src/update_prs.mjs'
 
 async function run() {
-  const octokit = {
-    rest: new Octokit({ auth: process.env.GITHUB_TOKEN }),
-    graphql: graphql.defaults({
-      headers: {
-        authorization: `token ${process.env.GITHUB_TOKEN}`,
-      },
-    }),
-  }
-  await generateReleaseNotes(octokit, {
-    owner: 'ExodusMovement',
-    repo: process.env.REPO || 'gh-release-generator-test',
-  })
+  const octokit = getOctokit()
+  const loginIdList = ['dooglio', 'nachoaIvarez', 'ChALkeR']
+
+  return await lookupLoginIds(octokit, loginIdList)
 }
 
-run().catch(console.error)
+run().then(console.info).catch(console.error)
